rag-chatbot/pyproject.toml
[project]
name = "rag-chatbot"
version = "0.1.0"
description = "RAG system with FastAPI, LangChain, hybrid retrieval, and observability"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn==0.30.6",
  "gunicorn==22.0.0",
  "pydantic-settings==2.4.0",
  "python-dotenv==1.0.1",
  "loguru==0.7.2",
  "prometheus-fastapi-instrumentator==7.0.0",
  "opentelemetry-sdk==1.27.0",
  "opentelemetry-instrumentation-fastapi==0.48b0",
  "redis==5.0.8",
  "langchain==0.2.12",
  "langchain-openai>=0.2.0",
  "langchain-community>=0.3.0",
  "qdrant-client==1.11.3",
  "chromadb==0.5.5",
  "pinecone-client==5.0.1",
  "rank-bm25==0.2.2",
  "pypdf==5.0.1",
  "unstructured==0.15.7",
  "markdown==3.6",
  "python-multipart==0.0.9",
  "sse-starlette==2.1.3",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.2",
  "pytest-asyncio==0.23.8",
  "httpx==0.27.0",
  "coverage==7.6.1",
  "mypy==1.11.2",
  "ruff==0.6.9",
  "black==24.8.0",
  "pre-commit==3.8.0",
  "locust==2.31.6",
]

[tool.black]
target-version = ["py311"]
line-length = 100

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "B", "UP", "N"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
namespace_packages = true

